requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
decorator==5.1.1
Django==4.1.7
djangorestframework==3.14.0
drf-spectacular==0.26.0
executing==1.2.0
inflection==0.5.1
ipdb==0.13.13
ipython==8.11.0
jedi==0.18.2
jsonschema==4.17.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
uritemplate==4.1.1
wcwidth==0.2.6
